+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **fix**: time not captured using exit
+
 ## 0.1.9
 
 - **fix**: ignore empty request data in companion